checkout the interface model and view files from sfu-1.8-stable
We are going to manages these files:

app/models/nic/interface.rb
app/views/nic/_base_form.html.erb

A problem was introduced when Foreman core introduced Nic models.

It revolves around how the dependency Nic::Managed#domain_id is
established. When looking for a domain attribute for a Host/Nic,
Foreman grabs the first match from Domains.all, so a host like...

"host-00.foo.bar.com" will match "bar.com"

... if it finds that first -- even if "foo.bar.com" exists!!!

What's more, this garbage domain value is stuck on the Host's
primary interface object (Nic::Base#domain_id) and when the
form view for the Host interfaces is rendered, Nic::Base#name
and Nic::Base#domain_id are synthesized into a non-authoritative
FQDN, like...

"host-00.foo"

This is wrong and it leads to a bunch of botched and sometimes
un-editable records.

We are going to patch this.

See the next two commits...