patch normalize_name method; fix domain select bug
Buried in this 'normalization' method is an setter routine.

Totally. Weird.

After actually normalizing Host.name, a Domain attribute for the primary interface (Nic::Managed) is set. It does this by scanning Domain.all, but it returns the _first_ match.

This is _not_ good and _not_ authoritative.

It's not even "best effort", it's ill-conceived and broken.

Fix this by returning an _exact_ match ONLY. No match found? Return nil.

Simple.