fix dangerous "DNS name" form value
This is part two of the changes required to make Foreman retain record names.

This tiny bit of view code was populating the DNS name form field with an abominable value using the Nic::Base#shortname method.

Change it Nic::Base#name so that it looks and feels like a real DNS name and it doesn't truncate the Host#name when you submit the form. Yay!