modify Host::Managed#fqdn
Right now, Smart Class Parameter overrides at the host level are broken. This is happening because Foreman is using the fqdn fact to perform lookups for a Host's Smart Class Parameter overrides.

This one small change is required to fix this.

Fix: Always return Host::Managed#name for Host::Managed#fqdn.

Prior to this, Host::Managed#fqdn method was returning the facts['fqdn'] first, then the record name.

We do not want this. Ever.

FQDNs can change. Facts get updated. Ugh. We cannot have that.

For us, the record name is the AUTHORITATIVE name -- never fqdn and although we are cheating a bit by returning the name instead of the fqdn, this will help stabilize Smart Class Parameter override lookups.